:recycle: 송이현 : 사이드바 interface 한파일로 통합
