+++ frontnext/src/app/ui/Insert/components/Sidebar/TableList.tsx
@@ -2,10 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { fetchTables } from '@/app/lib/fetchDataTest';
 import { useLanguage } from '../../../SettingMoules/LanguageContext';
 import common from '@/app/ui/styles/CommonStyles';
-
-interface TableListProps {
-  onTableClick: (tableName: string) => void;
-}
+import { TableListProps } from './components/SidebarProps';
 
 const TableList: React.FC<TableListProps> = ({ onTableClick }) => {
   const [tables, setTables] = useState<string[]>([]);
