--- frontnext/src/app/ui/Insert/components/Sidebar/components/ButtonProps.ts
@@ -1,4 +1,0 @@
-// components/ButtonProps.ts
-export interface AddButtonProps {
-  onClick: () => void;
-}
